requirements.txt
streamlit==1.32.0
streamlit-extras
streamlit-option-menu
openai-whisper
google-generativeai
transformers
torch
sentence-transformers
librosa
soundfile
pydub
moviepy
opencv-python
mediapipe
ffmpeg-python
imageio
imageio-ffmpeg
Pillow
python-dotenv
scipy
numpy
groq
cloudinary
